Refactor into logical groupings, fix random load order
needed to pause until all plugins are loaded before attempting to load custom order, otherwise, startup order will grab the order at a random time and load only the matching tabs that are currently loaded, the rest are then appended, creating an undesired random order
